+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+
+
+### Unreleased
+
+
+### Features
+
+* Added option parameter to getKey function
+
+
+
 ### [1.7.3](https://github.com/Kikobeats/cacheable-response/compare/v1.7.2...v1.7.3) (2019-09-21)
 
 ### [1.7.2](https://github.com/Kikobeats/cacheable-response/compare/v1.7.1...v1.7.2) (2019-09-11)